x64: ip: use kernels/blocking that reduce bandwidth
For small minibatch dimension, using blocking and kernels that reduce
bandwidth requirement of B matrix improves performance.  If minibatch
dimension is tiny we don't save much bandwidth if we only have one
m-loop iteration in the kernel and less FMAs will be executed.

Some cases with parallel reduction can be further improved to use more
threads to avoid tail cases that overload a thread.  Also, it is not
beneficial to reduce number of reduction threads if non-reduction
dimensions will still use single thread.